+++ exporters/otlp/include/opentelemetry/exporters/otlp/otlp_grpc_log_record_exporter_factory.h
@@ -23,12 +23,12 @@ public:
   /**
    * Create a OtlpGrpcLogExporter.
    */
-  static std::unique_ptr<opentelemetry::sdk::logs::LogExporter> Create();
+  static std::unique_ptr<opentelemetry::sdk::logs::LogRecordExporter> Create();
 
   /**
    * Create a OtlpGrpcLogExporter.
    */
-  static std::unique_ptr<opentelemetry::sdk::logs::LogExporter> Create(
+  static std::unique_ptr<opentelemetry::sdk::logs::LogRecordExporter> Create(
       const OtlpGrpcExporterOptions &options);
 };